Don't stop the fabmo daemon when running in Docker, since there's no systemd to ask that of

+++ build_scripts/stop_fabmo.sh
@@ -1,3 +1,8 @@
 #!/bin/sh
 
+if [ ! -f /.dockerenv ]; then
+# When running inside docker, systemd is hard to do right, so just skip it
+
 sudo systemctl stop fabmo
+
+fi
